requirements.txt
flask 
python-pptx
gunicorn
#asposeslidescloud
aspose.slides
